Move manga2 to original package
