+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/track/TrackAdapter.kt
@@ -1,0 +1,44 @@
+package eu.kanade.tachiyomi.ui.manga.track
+
+import android.support.v7.widget.RecyclerView
+import android.view.ViewGroup
+import eu.kanade.tachiyomi.R
+import eu.kanade.tachiyomi.util.inflate
+
+class TrackAdapter(controller: TrackController) : RecyclerView.Adapter<TrackHolder>() {
+
+    var items = emptyList<TrackItem>()
+        set(value) {
+            if (field !== value) {
+                field = value
+                notifyDataSetChanged()
+            }
+        }
+
+    val rowClickListener: OnRowClickListener = controller
+
+    fun getItem(index: Int): TrackItem? {
+        return items.getOrNull(index)
+    }
+
+    override fun getItemCount(): Int {
+        return items.size
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackHolder {
+        val view = parent.inflate(R.layout.item_track)
+        return TrackHolder(view, this)
+    }
+
+    override fun onBindViewHolder(holder: TrackHolder, position: Int) {
+        holder.bind(items[position])
+    }
+
+    interface OnRowClickListener {
+        fun onTitleClick(position: Int)
+        fun onStatusClick(position: Int)
+        fun onChaptersClick(position: Int)
+        fun onScoreClick(position: Int)
+    }
+
+}
