+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/track/TrackItem.kt
@@ -1,0 +1,6 @@
+package eu.kanade.tachiyomi.ui.manga.track
+
+import eu.kanade.tachiyomi.data.database.models.Track
+import eu.kanade.tachiyomi.data.track.TrackService
+
+data class TrackItem(val track: Track?, val service: TrackService)
