--- app/src/main/java/eu/kanade/tachiyomi/ui/manga2/MangaEvent.kt
@@ -1,5 +1,0 @@
-package eu.kanade.tachiyomi.ui.manga2
-
-import eu.kanade.tachiyomi.data.database.models.Manga
-
-class MangaEvent(val manga: Manga)
