--- app/src/main/java/eu/kanade/tachiyomi/ui/manga2/info/ChapterCountEvent.kt
@@ -1,16 +1,0 @@
-package eu.kanade.tachiyomi.ui.manga2.info
-
-import rx.Observable
-import rx.subjects.BehaviorSubject
-
-class ChapterCountEvent {
-
-    private val subject = BehaviorSubject.create<Int>()
-
-    val observable: Observable<Int>
-        get() = subject
-
-    fun emit(count: Int) {
-        subject.onNext(count)
-    }
-}
