--- app/src/main/java/eu/kanade/tachiyomi/ui/manga2/info/MangaFavoriteEvent.kt
@@ -1,16 +1,0 @@
-package eu.kanade.tachiyomi.ui.manga2.info
-
-import com.jakewharton.rxrelay.PublishRelay
-import rx.Observable
-
-class MangaFavoriteEvent {
-
-    private val subject = PublishRelay.create<Boolean>()
-
-    val observable: Observable<Boolean>
-        get() = subject
-
-    fun call(favorite: Boolean) {
-        subject.call(favorite)
-    }
-}
